feat: split FrameGeneral and FrameLast with interface
